Move the Activity class from lib/activity.py to core/activity.py
Also reindent it while at it.
